requirements.txt
pyyaml>=6.0.2,<7.0.0
pydantic>=2.10.3,<3.0.0
pyhumps>=3.8.0,<4.0.0
requests>=2.32.3,<3.0.0